requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
command-not-found==0.3
cryptography==2.3
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
dataset-loader==1.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.21ubuntu2
Django==2.2.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
entrypoints==0.3
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
httplib2==0.11.3
idna==2.6
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
length==1.0.1
louis==3.8.0
lxml==4.5.0
macaroonbakery==1.2.1
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
nmaxmin==1.0.1
notebook==6.0.3
numpy==1.18.1
oauthlib==2.1.0
olefile==0.46
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==4.7.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.32.0
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==1.8.4
python-dateutil==2.7.3
python-debian==0.1.34
python-slugify==4.0.0
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
pyzmq==19.0.1
reportlab==3.5.18
requests==2.21.0
requests-unixsocket==0.1.5
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
w3lib==1.21.0
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
